requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.0.4
djangorestframework==3.15.1
generics==6.0.0
gunicorn==21.2.0
idna==3.7
packaging==24.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
